# backend/application/src/main/resources/META-INF/openapi.yaml
openapi: 3.1.0
info:
  title: Surplasse API
  version: 0.1.0
  summary: The Surplasse REST API, defined contract-first.
  description: |
    Surplasse lets independent restaurants run their own direct ordering
    channel. This contract is the single source of truth of the API: the
    backend implements Java interfaces generated from it, the frontends
    consume a generated TypeScript client.

    Version 0.1 covers the public catalog, ordering, payment, restaurateur
    identity flows, the authenticated operational-order list and the
    order-intake control. Other Dashboard reads remain drafts (`x-draft`
    blocks, excluded from generation and compatibility checks) until their
    authorization is implemented end to end.

    Conventions: paths are prefixed with `/v1`, money amounts are integers
    in cents with a `Cents` suffix, timestamps are ISO 8601 UTC, and every
    error is an RFC 9457 Problem Details document.
  contact:
    name: Surplasse
    url: https://surplasse.com
servers:
  - url: /
    description: Current API origin, injected by the environment at runtime.
...
paths:
...
  /v1/payments:
    post:
      tags: [payment]
      operationId: createPayment
      summary: Create the payment session of an order
      description: |
        Creates (or returns, when replayed) the Stripe payment session of a
        `pending_payment` order: the backend re-checks product availability,
        recomputes the amount from the catalog, creates a direct-charge
        PaymentIntent on the establishment's connected account and returns
        its client secret and account identifier for the Payment Element. The browser
        confirmation is informative only: the signed Stripe webhook is the
        single source of truth that marks the order paid. An operational pause
        rejects new and replayed payment-session requests but never prevents a
        signed webhook from reconciling a Payment Intent already in flight.
      security:
        - tableSession: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentCreationRequest"
      responses:
        "201":
          description: The payment session, ready for the Payment Element.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentSession"
              example:
                id: b3a2c1d0-e9f8-4a7b-8c6d-5e4f3a2b1c0d
                orderId: 0d9e8c7b-6a5f-4e3d-8c2b-1a0f9e8d7c6b
                amountCents: 2250
                currency: EUR
                clientSecret: pi_3Nxxxxxxxxxxxx_secret_xxxxxxxxxxxxx
                connectedAccountId: acct_1Nxxxxxxxxxxxxxxxx
        "401":
          description: The table session token is missing, unknown or expired.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://surplasse.com/problems/table-session-expired
                title: Table session expired
                status: 401
                detail: The table session token is expired or unknown.
                instance: /v1/payments
        "404":
          description: Unknown order, or order outside this session.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://surplasse.com/problems/resource-not-found
                title: Resource not found
                status: 404
                detail: No order matches this identifier.
                instance: /v1/payments
        "409":
          description: Ordering was paused, the order is not payable anymore, or a product became unavailable.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                orderIntakePaused:
                  summary: New payments are paused
                  value:
                    type: https://surplasse.com/problems/order-intake-paused
                    title: Order intake paused
                    status: 409
                    detail: This establishment is not accepting new payments right now.
                    instance: /v1/payments
                orderNotModifiable:
                  summary: The order is already settled
                  value:
                    type: https://surplasse.com/problems/order-not-modifiable
                    title: Order not modifiable
                    status: 409
                    detail: Order 0d9e8c7b-6a5f-4e3d-8c2b-1a0f9e8d7c6b is already paid.
                    instance: /v1/payments
        "422":
          description: The establishment has no active Stripe Connect account able to accept charges.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://surplasse.com/problems/business-rule-violation
                title: Business rule violation
                status: 422
                detail: This establishment is not ready to accept payments.
                instance: /v1/payments
        "503":
          description: Stripe is unreachable after retries.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://surplasse.com/problems/dependency-unavailable
                title: Dependency unavailable
                status: 503
                detail: Stripe did not answer after retries.
                instance: /v1/payments
  /v1/refunds:
    post:
      tags: [payment]
      operationId: createRefund
      summary: Fully refund a paid order
      description: |
        Creates or replays the full Stripe refund of one order on behalf of
        authenticated establishment staff. The direct charge is refunded on
        the connected account frozen on the payment. When Surplasse collected
        an application fee, the same operation returns it to the restaurant.
        Partial refunds are outside the MVP.

        The request is idempotent. Replaying a key returns the exact refund it
        previously received, while a new key reuses any refund still in
        progress or already succeeded. A failed or canceled attempt can be
        retried with a new key. The order moves to `refunded` only after Stripe
        reports a successful refund, synchronously or through a signed webhook.
      security:
        - restaurateurSession: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefundCreationRequest"
      responses:
        "201":
          description: The full refund attempt, newly created or safely replayed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Refund"
              example:
                id: a4b3c2d1-e0f9-4876-95a4-b3c2d1e0f987
                orderId: 0d9e8c7b-6a5f-4e3d-8c2b-1a0f9e8d7c6b
                amountCents: 2250
                applicationFeeRefundedCents: 22
                currency: EUR
                reason: restaurant_refusal
                status: succeeded
        "401":
          description: The restaurateur session is missing or expired.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://surplasse.com/problems/session-expired
                title: Session expired
                status: 401
                detail: The restaurateur session is expired or missing.
                instance: /v1/refunds
        "404":
          description: Unknown order, order outside the caller's establishment, or no settled payment.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://surplasse.com/problems/resource-not-found
                title: Resource not found
                status: 404
                detail: No order matches this identifier.
                instance: /v1/refunds
        "409":
          description: The key conflicts, the order cannot be refunded, or a successful refund already exists.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://surplasse.com/problems/order-not-modifiable
                title: Order not modifiable
                status: 409
                detail: Order 0d9e8c7b-6a5f-4e3d-8c2b-1a0f9e8d7c6b cannot be refunded from status served.
                instance: /v1/refunds
        "422":
          description: Stripe rejects the refund for a business reason.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://surplasse.com/problems/business-rule-violation
                title: Business rule violation
                status: 422
                detail: Stripe rejected the refund for this order.
                instance: /v1/refunds
        "503":
          description: Stripe is unreachable after retries.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://surplasse.com/problems/dependency-unavailable
                title: Dependency unavailable
                status: 503
                detail: Stripe did not answer.
                instance: /v1/refunds
...
components:
  securitySchemes:
    tableSession:
      type: apiKey
      in: header
      name: X-Table-Session
      description: >
        Opaque anonymous table session token, delivered by
        `createTableSession` when a table QR is scanned. Bound to one
        establishment and one table, valid for the duration of a meal,
        carrying no personal data.
    restaurateurSession:
      type: apiKey
      in: cookie
      name: surplasse_session
      description: >
        Short session JWT set after the magic link exchange. The cookie is
        host-only for api.surplasse.com, HttpOnly, Secure in production,
        SameSite=Lax and scoped to Path=/. Used by Dashboard endpoints and SSE.
    restaurateurRefresh:
      type: apiKey
      in: cookie
      name: surplasse_refresh
      description: >
        Opaque rotating refresh token, hashed server side and valid for 30
        days. The host-only cookie is HttpOnly, Secure in production,
        SameSite=Lax and scoped to Path=/v1/auth/sessions.
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: >
        Client-generated UUID identifying this intention. Replaying the same
        request with the same key returns the original response without any
        duplicate; the same key with a different payload yields a 409
        `idempotency-key-conflict`.
      schema:
        type: string
        format: uuid
      example: 7f3a9c1e-4b2d-4e8a-9c5f-1d2e3f4a5b6c
...
  schemas:
    Problem:
      type: object
      description: >
        RFC 9457 Problem Details document, the single error format of the
        API. The `type` URI is stable and identifies the applicative error.
      properties:
        type:
          type: string
          description: Stable URI identifying the applicative error type.
          default: about:blank
        title:
          type: string
          description: Short, human-readable summary of the error type.
        status:
          type: integer
          description: HTTP status code of this occurrence.
          minimum: 100
          maximum: 599
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence.
        instance:
          type: string
          description: URI reference of the request that produced the error.
      required: [title, status]
      example:
        type: https://surplasse.com/problems/resource-not-found
        title: Resource not found
        status: 404
        detail: No establishment matches slug 'le-cormoran'.
        instance: /v1/establishments/le-cormoran/menu
...
    PaymentCreationRequest:
      type: object
      description: The order to open a Stripe payment session for.
      properties:
        orderId:
          type: string
          format: uuid
          description: Order to pay, at status pending_payment, inside the caller's table session.
      required: [orderId]
      example:
        orderId: 0d9e8c7b-6a5f-4e3d-8c2b-1a0f9e8d7c6b
    PaymentSession:
      type: object
      description: >
        A Stripe payment session for one order. The client secret feeds the
        Payment Element; the amount is recomputed server side.
      properties:
        id:
          type: string
          format: uuid
          description: Identifier of the payment attempt.
        orderId:
          type: string
          format: uuid
          description: Order being paid.
        amountCents:
          type: integer
          description: Amount to pay in cents, recomputed server side.
          minimum: 1
        currency:
          type: string
          description: ISO 4217 currency code.
        clientSecret:
          type: string
          description: Stripe PaymentIntent client secret, consumed by the Payment Element.
        connectedAccountId:
          type: string
          description: Stripe Connect account used by both the PaymentIntent and Stripe.js.
      required: [id, orderId, amountCents, currency, clientSecret, connectedAccountId]
      example:
        id: b3a2c1d0-e9f8-4a7b-8c6d-5e4f3a2b1c0d
        orderId: 0d9e8c7b-6a5f-4e3d-8c2b-1a0f9e8d7c6b
        amountCents: 2250
        currency: EUR
        clientSecret: pi_3Nxxxxxxxxxxxx_secret_xxxxxxxxxxxxx
        connectedAccountId: acct_1Nxxxxxxxxxxxxxxxx
    RefundCreationRequest:
      type: object
      description: The paid order and operational reason for its full refund.
      properties:
        orderId:
          type: string
          format: uuid
          description: Paid order owned by the authenticated establishment.
        reason:
          type: string
          description: Operational reason recorded by Surplasse and sent to Stripe metadata.
          enum: [restaurant_refusal, item_unavailable, service_incident]
      required: [orderId, reason]
      example:
        orderId: 0d9e8c7b-6a5f-4e3d-8c2b-1a0f9e8d7c6b
        reason: restaurant_refusal
    Refund:
      type: object
      description: A full refund attempt and its latest Stripe status.
      properties:
        id:
          type: string
          format: uuid
          description: Surplasse refund identifier.
        orderId:
          type: string
          format: uuid
          description: Refunded order.
        amountCents:
          type: integer
          minimum: 1
          description: Full customer amount refunded in cents.
        applicationFeeRefundedCents:
          type: integer
          minimum: 0
          description: Surplasse application fee returned to the restaurant in cents.
        currency:
          type: string
          description: ISO 4217 currency code.
        reason:
          type: string
          enum: [restaurant_refusal, item_unavailable, service_incident]
          description: Operational reason selected by establishment staff.
        status:
          type: string
          enum: [pending, requires_action, succeeded, failed, canceled]
          description: Latest authoritative Stripe refund status.
      required: [id, orderId, amountCents, applicationFeeRefundedCents, currency, reason, status]
      example:
        id: a4b3c2d1-e0f9-4876-95a4-b3c2d1e0f987
        orderId: 0d9e8c7b-6a5f-4e3d-8c2b-1a0f9e8d7c6b
        amountCents: 2250
        applicationFeeRefundedCents: 22
        currency: EUR
        reason: restaurant_refusal
        status: succeeded
...